+++ app/Http/Controllers/OrderController.php
@@ -73,13 +73,18 @@ class OrderController extends Controller
 
     public function destroy(Order $order)
     {
-        $order->delete();
-        return redirect()->route('order-index', ['#' . $order->id]);
+        if (!$order->rest_Order()->count()) {
+            $order->delete();
+            return redirect()->route('order-index', ['#' . $order->id])->with('ok', 'Order was delet');
+        } else {
+            $countOrder = $order->rest_Order()->count();
+            return redirect()->back()->with('not', 'Can,t delet order. User have open ' . $countOrder . ' offer. ');
+        }
     }
 
     public function status(Request $request, Order $order)
     {
-        $to = User::find($order->user_id);
+        $to = User::find($order->id);
         if ($order->status == 2) {
             // Mail::to($to)->send(new OrderCompleted($order));
             $order->status = 3;
@@ -93,10 +98,7 @@ class OrderController extends Controller
                 return $food;
             });
         $order->ticket = $request->ticket;
-
-        return view('back.orders.index', [
-            'orders' => $orders
-        ]);
+        return redirect()->route('order-index', ['#' . $order->id]);
     }
 
     public function shiped(Request $request, Order $order)
